third_party/WebKit/LayoutTests/fast/css/clear-activechain-list-expected.txt

Text in a span

Chain of active elements should be cleared, even if style :active sets display: none on the current active element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS background is "rgb(0, 255, 0)"
PASS color is "rgb(255, 255, 255)"
PASS successfullyParsed is true

TEST COMPLETE
